[in-progress] working through API update.

+++ src/lib/twz/include/twz/sys/key.h
@@ -1,12 +1,11 @@
 #pragma once
-
-#include <twz/_obj.h>
+#include <stdint.h>
 
 #define TWZ_KEY_PRI 1
 
 struct key_hdr {
 	unsigned char *keydata;
-	size_t keydatalen;
+	uint64_t keydatalen;
 	uint32_t type;
 	uint32_t flags;
 };
